+++ ErsatzTV.Application/Configuration/Commands/UpdateXmltvSettingsHandler.cs
@@ -16,11 +16,26 @@ public class UpdateXmltvSettingsHandler(
 {
     public async Task<Either<BaseError, Unit>> Handle(
         UpdateXmltvSettings request,
-        CancellationToken cancellationToken) => await ApplyUpdate(request.XmltvSettings);
+        CancellationToken cancellationToken)
+    {
+        int playoutDaysToBuild =
+            await configElementRepository
+                .GetValue<int>(ConfigElementKey.PlayoutDaysToBuild)
+                .IfNoneAsync(2);
+
+        if (playoutDaysToBuild < request.XmltvSettings.DaysToBuild)
+        {
+            return BaseError.New(
+                $"XMLTV days to build ({request.XmltvSettings.DaysToBuild}) cannot be greater than Playout days to build ({playoutDaysToBuild})");
+        }
+
+        return await ApplyUpdate(request.XmltvSettings);
+    }
 
     private async Task<Unit> ApplyUpdate(XmltvSettingsViewModel xmltvSettings)
     {
         await configElementRepository.Upsert(ConfigElementKey.XmltvTimeZone, xmltvSettings.TimeZone);
+        await configElementRepository.Upsert(ConfigElementKey.XmltvDaysToBuild, xmltvSettings.DaysToBuild);
 
         await using TvContext dbContext = await dbContextFactory.CreateDbContextAsync();